+++ custom_modules/biorobots.h
@@ -1,5 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
 /*
 ###############################################################################
 # If you use PhysiCell in your project, please cite PhysiCell and the version #
@@ -66,121 +64,53 @@
 #                                                                             #
 ###############################################################################
 */
---> 
 
-<!--
-<user_details />
--->
+#include "../core/PhysiCell.h"
+#include "../modules/PhysiCell_standard_modules.h" 
 
-<PhysiCell_settings version="devel-version">
-	<domain>
-		<x_min>-1000</x_min>
-		<x_max>1000</x_max>
-		<y_min>-1000</y_min>
-		<y_max>1000</y_max>
-		<z_min>-10</z_min>
-		<z_max>10</z_max>
-		<dx>20</dx>
-		<dy>20</dy>
-		<dz>20</dz>
-		<use_2D>true</use_2D>
-	</domain>
-	
-	<overall>
-		<max_time units="min">2880</max_time> <!-- 2 days * 24 h * 60 min -->
-		<time_units>min</time_units>
-		<space_units>micron</space_units>
-	
-		<dt_diffusion units="min">0.01</dt_diffusion>
-		<dt_mechanics units="min">0.1</dt_mechanics>
-		<dt_phenotype units="min">6</dt_phenotype>	
-	</overall>
-	
-	<parallel>
-		<omp_num_threads>4</omp_num_threads>
-	</parallel> 
-	
-	<save>
-		<folder>output</folder> <!-- use . for root --> 
+using namespace BioFVM;
+using namespace PhysiCell;
 
-		<full_data>
-			<interval units="min">2</interval>
-			<enable>true</enable>
-		</full_data>
-		
-		<SVG>
-			<interval units="min">2</interval>
-			<enable>true</enable>
-		</SVG>
-		
-		<legacy_data>
-			<enable>false</enable>
-		</legacy_data>
-	</save>
-	
-	<microenvironment_setup>
-		<variable name="director signal" units="dimensionless" ID="0">
-			<physical_parameter_set>
-				<diffusion_coefficient units="micron^2/min">1000</diffusion_coefficient>
-				<decay_rate units="1/min">.1</decay_rate> <!-- 1 for cells  --> 
-			</physical_parameter_set>
-			<initial_condition units="dimensionless">0</initial_condition>
-			<Dirichlet_boundary_condition units="dimensionless" enabled="false">1</Dirichlet_boundary_condition>
-		</variable>
-		
-		<variable name="cargo signal" units="dimensionless" ID="1">
-			<physical_parameter_set>
-				<diffusion_coefficient units="micron^2/min">1000</diffusion_coefficient>
-				<decay_rate units="1/min">.4</decay_rate> <!-- 1 for cells  --> 
-			</physical_parameter_set>
-			<initial_condition units="dimensionless">0</initial_condition>
-			<Dirichlet_boundary_condition units="dimensionless" enabled="false">1</Dirichlet_boundary_condition>
-		</variable>
-		
-		<options>
-			<calculate_gradients>true</calculate_gradients>
-			<track_internalized_substrates_in_each_agent>false</track_internalized_substrates_in_each_agent>
-			<!-- not yet supported --> 
-			<initial_condition type="matlab" enabled="false">
-				<filename>./config/initial.mat</filename>
-			</initial_condition>
-			<!-- not yet supported --> 
-			<dirichlet_nodes type="matlab" enabled="false">
-				<filename>./config/dirichlet.mat</filename>
-			</dirichlet_nodes>
-		</options>
-	</microenvironment_setup>		
-	
-	<user_parameters>
-		<random_seed type="int" units="dimensionless">0</random_seed> 
+// declare the cell types 
 
-		<!-- for microenvironment setup --> 
-		<cargo_signal_D type="double" units="micron/min^2">1e3</cargo_signal_D>
-		<cargo_signal_decay type="double" units="1/min">.4</cargo_signal_decay> <!-- 50 micron length scale -->
-		<director_signal_D type="double" units="micron/min^2">1e3</director_signal_D>
-		<director_signal_decay type="double" units="1/min">.1</director_signal_decay> <!-- 100 micron length scale -->
-		
-		<!-- for cell definitions -->
-		<elastic_coefficient type="double" units="1/min">0.05</elastic_coefficient>
+static Cell_Definition worker_cell; 
+static Cell_Definition cargo_cell; 
+static Cell_Definition director_cell; 
+static Cell_Definition linker_cell; 
 
-		<worker_motility_persistence_time type="double" units="min">5.0</worker_motility_persistence_time>
-		<worker_migration_speed type="double" units="micron/min">5.0</worker_migration_speed>
-		<attached_worker_migration_bias type="double" units="dimensionless">1.0</attached_worker_migration_bias>
-		<unattached_worker_migration_bias type="double" units="dimensionless">0.5</unattached_worker_migration_bias>
+static int worker_ID = 0;
+static int cargo_ID = 1;
+static int linker_ID = 2; 
+static int director_ID = 3;
 
-		<!-- tissue setup --> 
-		<number_of_directors type="int" units="none">15</number_of_directors> 
-		<number_of_cargo_clusters type="int" units="none">100</number_of_cargo_clusters>
-		<number_of_workers type="int" units="none">50</number_of_workers>
-		
-		<!-- cargo parameters --> 
-		<drop_threshold type="double" units="dimensionless">0.4</drop_threshold> 
-		
-		<!-- colors --> 
-		<worker_color type="string" units="none">red</worker_color>
-		<cargo_color type="string" units="none">blue</cargo_color>
-		<director_color type="string" units="none">limegreen</director_color> 
-	
-	</user_parameters>
-	
-</PhysiCell_settings>
+// set up the microenvironment 
+
+void setup_microenvironment( void ); // done 
+
+// set up the cell types 
+
+void create_cell_types( void );
+
+// set up the problem geometry 
+
+void setup_tissue( void ); 
+
+// coloring functions 
+
+std::vector<std::string> robot_coloring_function( Cell* pCell ); 
+
+// these are the custom functions for these cells 
+
+void extra_elastic_attachment_mechanics( Cell* pCell, Phenotype& phenotype, double dt );
+
+void worker_cell_rule( Cell* pCell, Phenotype& phenotype, double dt ); 
+void worker_cell_motility( Cell* pCell, Phenotype& phenotype, double dt ); 
+
+void cargo_cell_rule( Cell* pCell , Phenotype& phenotype , double dt ); 
+
+
+
+void director_cell_rule( Cell* pCell , Phenotype& phenotype , double dt );  // done 
+
+
+